Security hardening: fix 12 findings across C++, scripts, workflows
C++ fixes (iccanalyzer-lite):
- C3 (HIGH): Replace 4 unaligned reinterpret_cast<uint32_t*> with memcpy()
- C8 (HIGH): Add std::nothrow to 5 new CIccProfile allocations + nullptr checks
- C5/C6 (MEDIUM): Fix signed shift UB in read32()/Read32() via uint32_t casts
- C2 (MEDIUM): Fix icRealloc contract - don't free ptr on realloc failure
- C10 (MEDIUM): Wrap std::stoi in safe_stoi() with try/catch fallback
- C4 (MEDIUM): Add zero-byte file check before malloc in ninja mode
- C12 (LOW): Initialize cwd buffer to zero
- C19 (MEDIUM): Add depth limit (64) to RecursiveScan to prevent stack overflow

Shell script fixes:
- S2/S3 (MEDIUM): Fix argument parsing in ramdisk-sync-to-disk.sh and
  ramdisk-teardown.sh (for loop with $2 → while/shift pattern)

Workflow fixes:
- W1 (HIGH): Pin actions/cache and microsoft/setup-msbuild to commit SHA
- W3 (MEDIUM): Add top-level permissions: {} to codeql and gpt-5.2-codex workflows

+++ iccanalyzer-lite/IccTagParsers.h
@@ -111,7 +111,7 @@ inline const char* GetTagXMLName(uint32_t sig) {
 }
 
 inline uint32_t Read32(const unsigned char* data) {
-  return (data[0] << 24) | (data[1] << 16) | (data[2] << 8) | data[3];
+  return ((uint32_t)data[0] << 24) | ((uint32_t)data[1] << 16) | ((uint32_t)data[2] << 8) | (uint32_t)data[3];
 }
 
 inline uint16_t Read16(const unsigned char* data) {
